Remove all trace of GRDB
