--- Pendulum.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,14 +1,0 @@
-{
-  "pins" : [
-    {
-      "identity" : "grdb.swift",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/groue/GRDB.swift.git",
-      "state" : {
-        "revision" : "3df9e8e5c10bd152f02d9aa349a21a0c527e6bda",
-        "version" : "6.2.0"
-      }
-    }
-  ],
-  "version" : 2
-}
